backend/requirement.txt
Django==5.0.13
Faker==37.0.2
mysqlclient==2.2.7
djangorestframework==3.15.2